fix(surf-chat): fix npe

+++ surf-chat/src/main/java/dev/slne/surf/social/chat/provider/ChannelProvider.java
@@ -19,6 +19,10 @@ public class ChannelProvider {
   public void handleQuit(Player player) {
     Channel channel = Channel.getChannel(player);
 
+    if(channel == null) {
+      return;
+    }
+
     if(channel.getOwner().equals(player)) {
       channel.delete();
     }
